#12 :recycle: Split DictionaryCreator and its tests into multiple files to modularize it
